+++ pmm/mysql_metrics.go
@@ -84,7 +84,7 @@ func (a *Admin) AddMySQLMetrics(info map[string]string, mf MySQLFlags) error {
 		Address: a.Config.ClientAddress,
 		Service: &srv,
 	}
-	if _, err := a.consulapi.Catalog().Register(&reg, nil); err != nil {
+	if _, err := a.consulAPI.Catalog().Register(&reg, nil); err != nil {
 		return err
 	}
 
@@ -103,12 +103,12 @@ func (a *Admin) AddMySQLMetrics(info map[string]string, mf MySQLFlags) error {
 		// Add info to Consul KV.
 		d := &consul.KVPair{Key: fmt.Sprintf("%s/%s/%s", a.Config.ClientName, serviceID, o),
 			Value: []byte("OFF")}
-		a.consulapi.KV().Put(d, nil)
+		a.consulAPI.KV().Put(d, nil)
 	}
 
 	d := &consul.KVPair{Key: fmt.Sprintf("%s/%s/dsn", a.Config.ClientName, serviceID),
 		Value: []byte(info["safe_dsn"])}
-	a.consulapi.KV().Put(d, nil)
+	a.consulAPI.KV().Put(d, nil)
 
 	// Install and start service via platform service manager.
 	svcConfig := &service.Config{
@@ -142,12 +142,12 @@ func (a *Admin) RemoveMySQLMetrics() error {
 		Node:      a.Config.ClientName,
 		ServiceID: consulSvc.ID,
 	}
-	if _, err := a.consulapi.Catalog().Deregister(&dereg, nil); err != nil {
+	if _, err := a.consulAPI.Catalog().Deregister(&dereg, nil); err != nil {
 		return err
 	}
 
 	prefix := fmt.Sprintf("%s/%s/", a.Config.ClientName, consulSvc.ID)
-	a.consulapi.KV().DeleteTree(prefix, nil)
+	a.consulAPI.KV().DeleteTree(prefix, nil)
 
 	// Stop and uninstall service.
 	if err := uninstallService(fmt.Sprintf("pmm-mysql-metrics-%d", consulSvc.Port)); err != nil {